+++ internal/cmd/project/describe.go
@@ -1,0 +1,127 @@
+// Copyright Mia srl
+// SPDX-License-Identifier: Apache-2.0
+//
+// Licensed under the Apache License, Version 2.0 (the "License");
+// you may not use this file except in compliance with the License.
+// You may obtain a copy of the License at
+//
+//     http://www.apache.org/licenses/LICENSE-2.0
+//
+// Unless required by applicable law or agreed to in writing, software
+// distributed under the License is distributed on an "AS IS" BASIS,
+// WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+// See the License for the specific language governing permissions and
+// limitations under the License.
+
+package project
+
+import (
+	"context"
+	"fmt"
+	"io"
+	"net/url"
+
+	"github.com/mia-platform/miactl/internal/client"
+	"github.com/mia-platform/miactl/internal/clioptions"
+	"github.com/mia-platform/miactl/internal/encoding"
+	"github.com/spf13/cobra"
+)
+
+const (
+	describeProjectCmdUsage = "describe"
+	describeProjectCmdShort = "Describe a Project configuration"
+	describeProjectCmdLong  = `Describe the configuration of the specified Project.`
+)
+
+type describeProjectOptions struct {
+	ProjectID    string
+	RevisionName string
+	VersionName  string
+	OutputFormat string
+}
+
+// DescribeCmd returns a cobra command for describing a project configuration
+func DescribeCmd(options *clioptions.CLIOptions) *cobra.Command {
+	cmd := &cobra.Command{
+		Use:   describeProjectCmdUsage,
+		Short: describeProjectCmdShort,
+		Long:  describeProjectCmdLong,
+		RunE: func(cmd *cobra.Command, _ []string) error {
+			restConfig, err := options.ToRESTConfig()
+			cobra.CheckErr(err)
+
+			client, err := client.APIClientForConfig(restConfig)
+			cobra.CheckErr(err)
+
+			cmdOptions := describeProjectOptions{
+				RevisionName: options.Revision,
+				VersionName:  options.Version,
+				ProjectID:    restConfig.ProjectID,
+				OutputFormat: options.OutputFormat,
+			}
+
+			return describeProject(cmd.Context(), client, cmdOptions, cmd.ErrOrStderr())
+		},
+	}
+
+	flags := cmd.Flags()
+	options.AddProjectFlags(flags)
+	options.AddRevisionFlags(flags)
+	options.AddVersionFlags(flags)
+	options.AddOutputFormatFlag(flags, "json")
+
+	return cmd
+}
+
+func describeProject(ctx context.Context, client *client.APIClient, options describeProjectOptions, writer io.Writer) error {
+	if len(options.ProjectID) == 0 {
+		return fmt.Errorf("missing project name, please provide a project name as argument")
+	}
+
+	ref, err := getConfigRef(options.RevisionName, options.VersionName)
+	if err != nil {
+		return err
+	}
+
+	endpoint := fmt.Sprintf("/api/backend/projects/%s/%s/configuration", options.ProjectID, ref)
+	response, err := client.
+		Get().
+		APIPath(endpoint).
+		Do(ctx)
+	if err != nil {
+		return fmt.Errorf("failed to get project %s, ref %s: %w", options.ProjectID, ref, err)
+	}
+	if err := response.Error(); err != nil {
+		return err
+	}
+
+	projectConfig := make(map[string]any, 0)
+	if err := response.ParseResponse(&projectConfig); err != nil {
+		return fmt.Errorf("cannot parse project configuration: %w", err)
+	}
+
+	bytes, err := encoding.MarshalData(projectConfig, options.OutputFormat, encoding.MarshalOptions{Indent: true})
+	if err != nil {
+		return err
+	}
+
+	fmt.Fprintln(writer, string(bytes))
+	return nil
+}
+
+func getConfigRef(revisionName, versionName string) (string, error) {
+	if len(revisionName) > 0 && len(versionName) > 0 {
+		return "", fmt.Errorf("both revision and version specified, please provide only one")
+	}
+
+	if len(revisionName) > 0 {
+		encodedRevisionName := url.PathEscape(revisionName)
+		return fmt.Sprintf("revisions/%s", encodedRevisionName), nil
+	}
+	if len(versionName) > 0 {
+		encodedVersionName := url.PathEscape(versionName)
+		return fmt.Sprintf("versions/%s", encodedVersionName), nil
+	}
+
+	return "", fmt.Errorf("missing revision/version name, please provide one as argument")
+}